Fix #29, #31, #37, #38
